Create a dedicated PlotToolButton module
